Separate ArgFileIterator and write tests
